% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_records.R
\name{HI_record}
\alias{HI_record}
\title{Generate One or several Header Information (HI) records}
\usage{
HI_record(
  Country,
  Year,
  SeasonType,
  Season,
  Fleet,
  AreaType,
  FishingArea,
  DepthRange = NA,
  UnitEffort,
  Effort,
  AreaQualifier = NA
)
}
\arguments{
\item{Country}{character vector (mandatory)}

\item{Year}{numeric vector (mandatory)}

\item{SeasonType}{character vector. Keywords defined by ICES see doc (mandatory)}

\item{Season}{numeric vector (mandatory)}

\item{Fleet}{character vector. Keywords defined by ICES see doc (mandatory)}

\item{AreaType}{character vector. Keywords defined by ICES see doc (mandatory)}

\item{FishingArea}{character vector. Keywords defined by ICES see doc (mandatory)}

\item{DepthRange}{currently not used. Set to NA by default (optional)}

\item{UnitEffort}{character vector. Keywords defined by ICES see doc (optional)}

\item{Effort}{numeric vector (optional)}

\item{AreaQualifier}{currently not used. Set to NA by default (optional)}
}
\value{
A comma separated text string in InterCatch HI format
}
\description{
\code{HI_record} returns comma separated text line(s) to be used in a files imported into
ICES InterCatch database. See  \url{http://datsu.ices.dk/web/selRep.aspx?Dataset=76}.
To reduce the risk of enter data in the wrong place we recommend that you always use
named parameter list when calling this function
}
\examples{
 HI_record(Country = "SE", Year = 2019, SeasonType = "Month", Season = 6, Fleet = "FYK_C",
 AreaType = "SubDiv", FishingArea = "27.d.3.31", UnitEffort = "gd", Effort = NA)

}
