% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-ICimportr.R
\docType{package}
\name{ICimportr}
\alias{ICimportr}
\title{ICimportr: A package helping in the tedious to create InterCatch import fiules}
\description{
The foo package provides ...
}
\section{ICimportr functions}{

The foo functions ...
}

