% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_records.R
\name{SI_record}
\alias{SI_record}
\title{Generate One or several Species Information (SI) records}
\usage{
SI_record(
  Country,
  Year,
  SeasonType,
  Season,
  Fleet,
  AreaType,
  FishingArea,
  DepthRange = NA,
  Species,
  Stock = NA,
  CatchCategory = "L",
  ReportingCategory = "R",
  DataToFrom = NA,
  Usage = "H",
  SamplesOrigin = NA,
  QualityFlag = NA,
  UnitCATON,
  CATON,
  OffLandings = -9,
  varCATON = -9,
  InfoFleet = NA,
  InfoStockCoordinator = NA,
  InfoGeneral = NA
)
}
\arguments{
\item{Country}{character vector (mandatory)}

\item{Year}{Season numeric vector (mandatory)}

\item{SeasonType}{character vector. Keywords defined by ICES see doc (mandatory)}

\item{Season}{Season numeric vector (mandatory)}

\item{Fleet}{character vector. Keywords defined by stock coordinator see doc (mandatory)}

\item{AreaType}{character vector. Keywords defined by ICES see doc (mandatory)}

\item{FishingArea}{character vector. Keywords defined by ICES see doc (mandatory)}

\item{DepthRange}{currently not used. Set to NA by default (optional)}

\item{Species}{character vector. Keywords defined by ICES see doc (mandatory)}

\item{Stock}{character vector. Keywords defined by ICES see doc (default NA)}

\item{CatchCategory}{character vector. Keywords defined by ICES see doc (default L)}

\item{ReportingCategory}{character vector. Keywords defined by ICES see doc (default R)}

\item{DataToFrom}{character vector. Keywords defined by ICES see doc (default NA)}

\item{Usage}{character vector. Keywords defined by ICES see doc (default H)}

\item{SamplesOrigin}{character vector. Keywords defined by ICES see doc (default NA)}

\item{QualityFlag}{character vector. Keywords defined by ICES see doc (default NA)}

\item{UnitCATON}{character vector. Keywords defined by ICES see doc (mandatory)}

\item{CATON}{numeric vector (mandatory)}

\item{OffLandings}{numeric vector (optional)}

\item{varCATON}{numeric vector (optional)}

\item{InfoFleet}{character vector (optional)}

\item{InfoStockCoordinator}{character vector (optional)}

\item{InfoGeneral}{character vector (optional)}
}
\value{
Comma separated text string(s) in InterCatch SI format
}
\description{
\code{SI_record} returns comma separated text line(s) to be used in a files imported into
ICES InterCatch database. See  \url{http://datsu.ices.dk/web/selRep.aspx?Dataset=76}.
To reduce the risk of enter data in the wrong place we recommend that you always use
named parameter list when calling this function
}
\examples{
SI_record(Country = "SE", Year = 2020, SeasonType = "Month", Season = 6,
 Fleet = "MIS_R", AreaType = "SubDiv", FishingArea = "27.d.3.29",
 Species = "TRS", UnitCATON = "kg", CATON = 16)

}
