% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-ICimportr.R
\docType{package}
\name{ICimportr}
\alias{ICimportr-package}
\alias{ICimportr}
\title{ICimportr: A package helping you in the tedious to create InterCatch import files}
\description{
The ICimportr package provides functions to format HI, SI and SD records according to the
InterCatch Exchange format. The package alse contains two data tables \code{fishing_area}
and \code{fleet_key}. The tables can be used to assign metiers according to gear type and
to set correct SubDiv-codes as required by ICES. The tables might be tuned for Swedish
conditions but I am open for suggestions.
}
\section{ICimportr functions}{

The ICimportr functions are \code{\link{HI_record}}, \code{\link{SI_record}} and
\code{\link{SD_record}}
}

\author{
\strong{Maintainer}: Anders Kagervall \email{anders.kagervall@slu.se} (\href{https://orcid.org/0000-0003-4790-2696}{ORCID})

}
